refactor emotions, emotion class is now parent class of playeremotion and enemyemotion class
